Move TestProjectRule to .test package too
