+++ core/src/test/groovy/com/novoda/gradle/release/JavaDifferentGradleVersions.groovy
@@ -1,6 +1,6 @@
 package com.novoda.gradle.release
 
-import com.novoda.gradle.release.rule.TestProjectRule
+import com.novoda.gradle.release.test.TestProjectRule
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 import org.junit.Rule
